+++ bash/sdl/install_sdl_mixer.sh
@@ -1,17 +1,26 @@
 #!/bin/bash
 
+disabler="--disable-music-mod --disable-music-midi --disable-music-ogg --disable-music-flac --disable-music-mp3 --disable-music-opus --disable-music-mod-modplug"
+
 echo " > installing [sdl_mixer] from source"
 # Default output parameter
 default_sdl_mixer_path="build/sdl_mixer" # linux-sdl_image
 # Check if an argument is provided, if not use the default
 sdl_mixer_path="${1:-$default_sdl_mixer_path}"
+lib_file_type="${2:-'so'}"
+if [ "$lib_file_type" = "dll" ]; then
+    lib_file_name="SDL2_mixer"
+else
+    lib_file_name="libSDL2_mixer"
+fi
+lib_file="$lib_file_name.$lib_file_type"
+echo " > sdl path is [$sdl_mixer_path] lib file [$lib_file]"
 
 sdl_mixer_url="https://github.com/libsdl-org/SDL_mixer/releases/download/release-2.8.0/SDL2_mixer-2.8.0.zip"
 sdl_mixer_zip="build/sdl_mixer.zip"
 sdl_mixer_source="$sdl_mixer_path/SDL2_mixer-2.8.0" # 2.0.4"
 sdl_mixer_include="include/sdl_mixer"
 sdl_mixer_include_source="$sdl_mixer_path/include"
-lib_file="libSDL2_mixer.so"
 sdl_mixer_lib_file="lib/$lib_file"
 sdl_mixer_lib_file_source="$sdl_mixer_path/build/.libs/$lib_file"
 sdl_mixer_lib_file2="lib/libSDL2_mixer-2.0.so.0"
@@ -38,26 +47,51 @@ if [ ! -f $sdl_mixer_lib_file ]; then
         echo " > lib file detected [$sdl_mixer_path/$lib_file]"
     else
         echo " + building [$lib_file]"
-        cd $sdl_mixer_path
-        echo " > cleaning source of sdl_mixer"
-        make clean > /dev/null 2>&1
-        echo " > configuring sdl_mixer"
-        ./autogen.sh > /dev/null 2>&1
-        ./configure --enable-shared --disable-static > /dev/null 2>&1
-        echo " > making sdl_mixer"
-        make > /dev/null 2>&1
-        cd ../..
+        if [ -f $sdl_mixer_lib_file_source ]; then
+            echo " > [$sdl_mixer_lib_file_source] already made"
+        else
+            cd $sdl_mixer_path
+            echo " > cleaning source of sdl_mixer"
+            make clean > /dev/null 2>&1
+            echo " > configuring sdl_mixer"
+            ./autogen.sh > /dev/null 2>&1
+            if [ "$lib_file_type" = "dll" ]; then
+                echo "   - configuring for mingw32"
+                CPPFLAGS="-I../../include/sdl" ./configure --host=x86_64-w64-mingw32 --enable-shared --disable-static $disabler --with-sdl2-prefix="../sdl" > /dev/null 2>&1
+                #  "../../lib/sdl-dll"
+            else
+                ./configure --enable-shared --disable-static $disabler --with-sdl-prefix="../../lib" > /dev/null 2>&1
+            fi
+            if [ $? -ne 0 ]; then
+                echo " ! sdl_mixer configure failed"
+                exit
+            else
+                echo " + configure successful"
+            fi
+            echo " > making sdl_mixer"
+            make # > /dev/null 2>&1
+            if [ $? -ne 0 ]; then
+                echo " ! sdl_mixer make failed"
+                exit
+            else
+                echo " + make successful"
+            fi
+            cd ../..
+        fi
     fi
     if [ ! -f $sdl_mixer_lib_file_source ]; then
         echo " ! failed to make [$sdl_mixer_lib_file_source]"
     else
-        echo " + copying to [$sdl_image_lib_file]"
+        echo " + copying to [$sdl_mixer_lib_file]"
         cp $sdl_mixer_lib_file_source $sdl_mixer_lib_file
     fi
 fi
-if [ ! -f $sdl_mixer_lib_file2 ]; then
-    echo " + copying to [$sdl_mixer_lib_file2]"
-    cp $sdl_mixer_lib_file_source2 $sdl_mixer_lib_file2
+
+if [ ! "$lib_file_type" == "dll" ]; then
+    if [ ! -f $sdl_mixer_lib_file2 ]; then
+        echo " + copying to [$sdl_mixer_lib_file2]"
+        cp $sdl_mixer_lib_file_source2 $sdl_mixer_lib_file2
+    fi
 fi
